+++ src/utils/omni-search-util/omni-search-util.spec.ts
@@ -228,7 +228,7 @@ describe("getItemsFromOmniValue", () => {
   it("eliminates empty comma sections", () => {
     expect(getItemsFromOmniValue(" 12 1, , 34")).toEqual(["12 1", "34"]);
   });
-  // TODO(jrosenfield): add more testing coverage
+  // TODO(jzhao): add more testing coverage
 });
 
 describe("getValueItemsFromSearchValues", () => {